+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for homepage and repository urls
 - Command to update the information of an already existing plugin
 - Command to update the logo of an already existing plugin
+- Command to update the store page of an already existing plugin
 
 ## 0.1.0 - 2023-09-01